Revise 7th principle of learning description
The 7th principle of learning contained a section that described instructional scaffolding specifically as a Parson's Problem. While a Parson's Problem is used in scaffolding instruction design, it is not scaffolding.  I changed the description to describe scaffolding, separating it from parson's problems, and emphasized that the use of instructional scaffolding in carpentry workshops is what makes them successful.