link to styles repo setup
Clarify that lesson developers need to create the lesson repo first using the instructions linked. This is a temporary fix for https://github.com/carpentries/curriculum-development/issues/26. Ideally, all of this documentation will be in one place, so migrating the setup instructions from lesson-example to the curriculum development handbook probably makes sense.